scripts/seeder/requirements/base.txt
Faker
psycopg2-binary
pydantic
pyyaml
python-ulid
bcrypt
ulid
minio
argon2
